updated long to Integer and added inventory dto

+++ order-inventory/src/main/java/com/spring/order_inventory/dto/ShipmentResponseDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ShipmentResponseDto {
 
-    private Long shipmentId;
+    private Integer shipmentId;
     private String shipmentStatus;
     private String deliveryAddress;
 }
